task/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
numpy==1.17.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
